Make updates to support executing immediate run commands and preserving needed metadata for the run to succeed (e.g., pid, downloaddir, seqNo, etc...)
