+++ cmd/immediateruncommandservice/main.go
@@ -1,22 +1,28 @@
 package main
 
 import (
-	"fmt"
 	"os"
 
-	"github.com/Azure/run-command-handler-linux/internal/immediatecmds"
-	"github.com/Azure/run-command-handler-linux/internal/runcommandcommon"
+	"github.com/Azure/run-command-handler-linux/internal/immediateruncommand"
+	"github.com/Azure/run-command-handler-linux/pkg/versionutil"
+	"github.com/go-kit/kit/log"
+)
+
+// These fields are populated by govvv at compile-time.
+var (
+	Version   string
+	GitCommit string
+	BuildDate string
+	GitState  string
 )
 
 // Entry Point for ImmediateRunCommand
 func main() {
-	cmd, ok := immediatecmds.Cmds["runService"]
+	// After starting the program, vars from versionutil.go must be set in order to share those values across the program.
+	versionutil.Initialize(Version, GitCommit, BuildDate, GitState)
 
-	// If no command was found, then exit
-	if !ok {
-		fmt.Printf("missing runService implementation")
-		os.Exit(2)
-	}
-
-	runcommandcommon.ProcessGoalState(cmd)
+	ctx := log.NewContext(log.NewSyncLogger(log.NewLogfmtLogger(
+		os.Stdout))).With("time", log.DefaultTimestamp).With("version", versionutil.VersionString())
+	ctx = ctx.With("operation", "runService")
+	immediateruncommand.StartImmediateRunCommand(ctx)
 }
